Extracted settings classes for the cli app to individual files instead of public in public
