+++ Src/WitsmlExplorer.Console/ListCommands/ListLogsCommand.cs
@@ -12,17 +12,8 @@ using WitsmlExplorer.Console.WitsmlClient;
 
 namespace WitsmlExplorer.Console.ListCommands
 {
-    public class ListLogsCommand : AsyncCommand<ListLogsCommand.ListLogsSettings>
+    public class ListLogsCommand : AsyncCommand<ListLogsSettings>
     {
-        public class ListLogsSettings : CommandSettings
-        {
-            [CommandArgument(0, "<WELL_UID>")]
-            public string WellUid { get; init; }
-
-            [CommandArgument(1, "<WELLBORE_UID>")]
-            public string WellboreUid { get; init; }
-        }
-
         private readonly IWitsmlClient witsmlClient;
 
         public ListLogsCommand(IWitsmlClientProvider witsmlClientProvider)
