+++ Src/WitsmlExplorer.Console/ListCommands/ListTubularsCommand.cs
@@ -12,17 +12,8 @@ using WitsmlExplorer.Console.WitsmlClient;
 
 namespace WitsmlExplorer.Console.ListCommands
 {
-    public class ListTubularsCommand : AsyncCommand<ListTubularsCommand.ListTubularsSettings>
+    public class ListTubularsCommand : AsyncCommand<ListTubularsSettings>
     {
-        public class ListTubularsSettings : CommandSettings
-        {
-            [CommandArgument(0, "<WELL_UID>")]
-            public string WellUid { get; init; }
-
-            [CommandArgument(1, "<WELLBORE_UID>")]
-            public string WellboreUid { get; init; }
-        }
-
         private readonly IWitsmlClient witsmlClient;
 
         public ListTubularsCommand(IWitsmlClientProvider witsmlClientProvider)
